Make sure deleteById does not contain null values
Unfortunately this library has to support old PHP versions as well, so
we are not able to use type-hints to prevent null values (for now). If
a user passes null values to certain deleteById methods this could
result in unwanted data loss, so we throw an exception to prevent any
loss.

+++ src/Requests/Application/Custom/ControllerRequests.php
@@ -3,6 +3,7 @@
 namespace RIPS\Connector\Requests\Application\Custom;
 
 use GuzzleHttp\RequestOptions;
+use RIPS\Connector\Exceptions\LibException;
 use RIPS\Connector\Requests\BaseRequest;
 
 class ControllerRequests extends BaseRequest
@@ -124,6 +125,10 @@ class ControllerRequests extends BaseRequest
      */
     public function deleteById($appId, $customId, $controllerId, array $queryParams = [])
     {
+        if (is_null($appId) || is_null($customId) || is_null($controllerId)) {
+            throw new LibException('appId, customId, or controllerId is null');
+        }
+
         $response = $this->client->delete($this->uri($appId, $customId, $controllerId), [
             'query' => $queryParams,
         ]);
