+++ src/Requests/Application/Custom/SourceRequests.php
@@ -3,6 +3,7 @@
 namespace RIPS\Connector\Requests\Application\Custom;
 
 use GuzzleHttp\RequestOptions;
+use RIPS\Connector\Exceptions\LibException;
 use RIPS\Connector\Requests\BaseRequest;
 
 class SourceRequests extends BaseRequest
@@ -124,6 +125,10 @@ class SourceRequests extends BaseRequest
      */
     public function deleteById($appId, $customId, $sourceId, array $queryParams = [])
     {
+        if (is_null($appId) || is_null($customId) || is_null($sourceId)) {
+            throw new LibException('appId, customId, or sourceId is null');
+        }
+
         $response = $this->client->delete($this->uri($appId, $customId, $sourceId), [
             'query' => $queryParams,
         ]);
